Conditionalize docker socket bind-mount
The socket is only needed when ContainerCli is set to 'docker'.
It only affects mistral executor and sensu-client containers, which were
the last containers relying on the socket.

For sensu-client, it was for healthchecks and they are being replaced by
systemd so the feature parity will be here.

For mistral-executor, it's needed by tripleo-validations running docker
CLI and they will have to run podman cli instead of docker.

Change-Id: I4e3d29a6eb65d871d7a1a935fcbd7bb98e7d1752