haproxy: deploy IPtables rules from the host
Calling iptables CLI in the container requires advanced and risky
bind-mounts, and on certain platform, iptables-save can't be found (e.g.
fedora28 containers).

This patch simplifies the firewall step for HAproxy container
configuration where we now run Puppet on the host instead of from the
container.

Note: we can't use the puppet module in Ansible yet because we need
Ansible 2.7.6 which has:
https://github.com/ansible/ansible/commit/8606fb33f0e8954b588eaecec2d99b4a120fd4ad
In the meantime, we use shell.


Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>
Co-Authored-By: Cédric Jeanneret <cjeanner@redhat.com>

Change-Id: Ia66db8e4ab0ccec7cc86665e2ad32d2861fe30c8