Enable Glance image_conversion plugin when backend is rbd
Ceph can store QCOW2 images but to be able to create Nova guests
from image snapshots the image format must be in RAW format,
otherwise Nova will download the QCOW2 image, convert it locally
and upload it back into the Ceph pool if the ephemeral storage is
also Ceph. This process slows down guest boot especially when
multiple guests are booted in parallel and consumes a lot of CPU
and network resources.

This change enables the Glance image_conversion plugin when the
Glance backend is set to 'rbd', which takes care of converting
the uploaded images into RAW format automatically when operators
use the glance 'import' command.

Change-Id: I9fc0758dbe64f3fcf0e990f4253151e9472a6ab6