Store nodes information in a dict
Re-using the same variable for each role was not very clean and forced
us to re-check which master nodes were new in the deployment. Instead,
use a dict for `role_nodes` and `new_role_nodes` with keys being the
role name.

This will allow us to re-generate updated inventory in later step
without needing to set the facts once again.

Related-Bug: #1812962
Change-Id: I21ed56a1c4d43a3e0945c8b74dd94ccf89dc05b0