Fix nova_cell_v2_discover_host.py with python3
When this script runs via python3 it fails with:

  "stdout: ",
  "stderr: + command -v python3",
  "+ python3 /docker-config-scripts/nova_cell_v2_discover_host.py",
  "Traceback (most recent call last):",
  " File \"/docker-config-scripts/nova_cell_v2_discover_host.py\", line 75, in <module>",
  " TypeError: a bytes-like object is required, not 'str'"

This is because in python3 subprocess.check_output() will return bytes
and trying to split it using a string '\n' will break with the error
above. Let's just use the universal_newlines=True parameter which we
have been using everywhere in tripleo so far.
Also skip any empty lines that might show up in the output which would
give the error:

  ValueError: not enough values to unpack (expected 2, got 0)

Tested with a python3 deployment and got a successful deployment (rhel8
os + f28 based-containers).

Change-Id: Ic7904c4f3027cc5e7e05d52757e36dbc05f3d487
Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>
Closes-Bug: #1813014