deployed-servers: symlink ansible-playbook
In RHEL8, we'll have /usr/bin/ansible-playbook provided by ansible package.
Everywhere in TripleO we rely on /usr/bin/ansible-playbook but the
ansible-python3 package in Fedora 28 doesn't provide it.

Until it does, we have a temporary workaround that will create a symlink only
if the file doesn't exist and if ansible-playbook3 is here.

Change-Id: I520bf0d665fa9a331fe94168595d3b09dbf1795c
Related-Bug: #1812837