[Queens only] Include new parameter RHELRegistrationUpgrade to rhel-registration.
Setting the rhel_reg_force parameter to true during upgrade prepare
doesn't work correctly when the customer passes the rhel-registration/
environment-rhel-registration.yaml, as this template sets an empty
value for that parameter by default (an which is usually not modified
by the user). As environments/lifecycle/upgrade-prepare.yaml is prepended,
any template setting a different value for rhel_reg_force param will have
precedence on what we set in the upgade-prepare.yaml env file.

The solution is creating a new parameter RHELRegistrationUpgrade, which
defaults to false and which we ensure it's set to true during upgrade
prepare by setting it in environments/lifecycle/upgrade-prepare.yaml.
When RHELRegistrationUpgrade is set to True we enforce that REG_FORCE
sets to True, otherwise it will simply take the value passed in the heat
parameter rhel_reg_force.

Closes-Bug: #1813184

Change-Id: Idc1bdad0a19150569dacb88abd539d86741bc643