[stable/queens] Add a check for kernels args update to avoid unnecessary reboot
In newton NFV deployments, the kernel args are applied using the
userdata scripts (first-boot). In queens NFV deployments, the userdata
scripts are rebooted, instead PreNetworkConfig resource is added
to allow kernel args configuration and reboot. During FFU, instead
of using the first-boot, migrating to PreNetworkConfig resource,
requires a check to ensure that a additional reboot is not applied.

After queens fresh deployment and subsequent updates, the reboot is
avoided by removing UPDATE in the deployment_actions of PreNetworkConfig.
During FFU, addition of PreNetworkConfig will result in the CREATE
action, as it is created first time (not available in newton), this
additional check is required to avoid reboot.
Closes-Bug: #1813610

Change-Id: I89c7a686c301861defe23963f6aa8c36c43c4bb5