Fix generation of configs that contain password files
In I8fe9a640ba36288a1f9cb18563b363159d4731c0 we added the ability
to prevent overwriting password files during docker-puppet runs, to
give the service the ability to update his own user credentials.

This doesn't work in case a stack update is running and config files
don't exist on the host in the first place (e.g. because of a
previous deploy failure, or due to a controller node replacement).
This also causes complications if a password file is already present
during a stack creation (e.g re-creating a stack on a split-stack
environment).

Change the way password files are handled:

  . if a previous password file exists on the host, do not overwrite
    it with the new password. Only use the new password for
    computing the hash.

  . otherwise, always copy the newly generated password file on the
    host.

Also, fix the config hash generation that currently considers the
password file twice, which makes the hash vary and cause
unexpected service restart at each stack update.

Change-Id: Ia77f1a82c4164f53fa90a6f05ba728787622285d
Closes-bug: #1809145