Do not pull image while tagging pcmk images in upgrade_tasks.
During the upgrad of pacemaker managed services there is
a temporary tagging of the existing container image to the
newly image we are going to upgrade to. The input during this
taggins is the container image id, in contrast to what is
passed during deploy tasks, which is the image name with its
tag. For that reason, we can't pull the image in these cases
otherwise the role will fail as it can't find an image to pull
from a container image id.

Change-Id: I4fdd3a05465fa0318b4ec5c079d59f4dd80fa9f2
Closes-Bug: #1814104