mysql: do not stop container when upgrade doesn't update mysql image
In 0015cc74416445c3cf4e18e566a954f609f7cf0f we added the ability
to stop gracefully the mysql container during a major upgrade, to
allow upgrade to versions of mysql with different redo log format [1].

Stopping mysql is OK as long as paunch restarts it during the deploy
tasks. However, paunch only (re)start mysql if the mysql config hash
changes, it doesn't check whether mysql is stopped or not. So running
"undercloud upgrade" with identical mysql container image now fails.

Fix the upgrade strategy by stopping mysql only when needed, that is
when mysql container image is to be updated.

[1] https://jira.mariadb.org/browse/MDEV-14848

Change-Id: I88f1913581e6c9e279bcaf99ad4260dd940636c3