Catch directories we can not change ownership
When there is a directory where the owner can not be changed, we
fail and do not proceed to check the remaining directories. This
has been seen where netapp creates a special .snapshot directory
which is protected and visible when useing NFSv3, not visible
when using NFSv4.

This change catches the error and proceeds.

Change-Id: I2e61030c863041cd73cb204e31e423da3c6f6944
Resolves-Bug: 1814260
(cherry picked from commit 898154857501b09a0101d4b7d8cbbd2c4c8b64d6)