placement: Add nova_api data extraction step during deployment
This change adds an additional deployment step that will attempt to
extract all Placement data from the nova_api database ahead of db syncs
being preformed. For the time being this is a noop as there should be no
data to move across. Eventually this will be used during upgrades and
actually used to migrate data between the nova_api and placement
database.

Co-Authored-By: Martin Schuppert <mschuppert@redhat.com>

Change-Id: Ifaa1101d05b835529730002ef985990c6469a449