Apply stop period for paunch-managed stateful svcs
Apply stop_grace_period of a 60s for stateful services managed via
paunch. That gives it time to attempt stopping it gracefully, while
applying configuration updates.

Also switch non-HA mysql template to run mysqld directly.
This is because myslqd_safe does not process the graceful stop
signals well, so we need to signal mysqld directly.
Note: mysqld_safe helps for better coverging HA clusters so
we leave it as is for the HA template.

Change-Id: I329f871b6f2509c0133823ca4c22238e3e19c76d
Related-bug: #1810690
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>