Do not restart bundles during a minor update
It makes no sense since we are going to stop the cluster on the node
anyway. With change Ib3562adbd83f7162c2aeb450329b7cc4ab200fc2 we
inject the TRIPLEO_MINOR_UDPATE variable into the update playbooks
and we also push it (by default to false) inside the paunch start
container step.

By exporting the variable TRIPLEO_MINOR_UPDATE without a value we
tell docker (and podman) to inherit this value from the host.
When the env variable is set to true we avoid calling pcs inside
the containers to restart the bundles because at that point we
know that restarting the HA container would make no sense.

We tested the changes as follows:
A) Ran a minor update (openstack overcloud update prepare + openstack
overcloud update run --roles Controller) and observed that there were no
extra pcmk-managed bundles restarts (i.e. even when paunch decided to
invoke the <service>_restart_bundle it was a noop because
TRIPLEO_MINOR_UDPATE was 'true')
B) We ran a re-deploy without any changes and observed that no restart
of pcmk-managed bundles took place
C) We ran a re-deploy and forced some configuration changes and observed
that TRIPLEO_MINOR_UDPATE was false and that the
<service>_restart_bundle was run by paunch and that it restarted the
bundle that had the config changed.

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>

Related-Bug: #1823849

Change-Id: Id1d671506d3ec827bc311b47d9363952e1239ce3
(cherry picked from commit e6ab4856d7982b208a16443dc3e7e0f3fd136113)