Add GlobalConfigExtraMapData
Adds a new GlobalConfigExtraMapData parameter that can be used to inject
global_config_settings hieradata into the deployment. Any values generated
in the stack will override those passed in by the parameter value.

This will be used for the distributed compute node when deploying with separate
stacks and data from the control plane stack needs to be injected into the
compute stack.

Change-Id: Id3e52e272bae67ee4036c81b3d7640255e0349ae