env/docker: add ContainerCli
This file will be used until we really get Podman everywhere (so when
CentOS 8 is out).
We still need Docker around for when Pacemaker is used. So this patch
will mainly be used by TripleO CI OVB jobs, so Docker can still be
deployed on the Overcloud, and not Podman.

blueprint podman-support

Change-Id: I80b07c83aae752b2b327a4570bc3df230f6a6319