Default ControlPlaneMtu to 1500
As a follow up to https://review.openstack.org/#/c/635622/, this also
defaults the value for mtu to 1500 when using DeployedServerPortMap.

Previously, the value was getting set to "0" since it did not get set in
the output. While the "0" value showed up in
/etc/os-net-config/config.json, it didn't actually break anything since
os-net-config has code that only set MTU with "if interface.mtu ...",
and 0 evaluated to False, so it assumed it wasn't set.

With this patch the value will be properly defaulted to 1500 though.

Change-Id: I12d110d653a77bef87d06b4713704844a46b065d