Added tar option to preserve metadata of existing fetch_dir
Added tar option to allow the current user to unpack local
or downloaded data (from swift) into ceph-ansible/fetch_dir,
preserving the existing metadata.
Added also a "/*" as trailing character on the *path field* of
the *archive* ansible task to create a tar.gz containing all
the fetch_dir content whithout including the parent directory
in the resulting path: in this way we should be able to avoid
nested fetch_dir.

1. Added a step to fix broken environment excluding the nested
fetch_dir from the archive sections;

2. Added a new task to delete from filesystem the nested
section to make everything clean: this task is executed
according to the output of the previous one (the archive task).

Change-Id: I45c394bd6b5c09730399202b476ed913e8f334c0
Related-Bug: 1816094