Only request neutron certificate from neutron dhcp service
The certificate request for the "neutron" certificate was set in the
neutron base template. This had the secondary effect of causing every
node that has a neutron service to try to request the certificate.

This fixes that issue by moving those bits to where the certificate is
actually used (which is only by the dhcp agent).

Change-Id: I10ade8a4b5ec30872210c633d35273309ae20377
Closes-Bug: #1816465