Backup /etc/os-net-config/config.json when os-net-config runs
Currently os-net-config is called by TripleO with either the
config.json or element_config.json configuration, and then
the configuration file is deleted after os-net-config runs.
This makes it more difficult to troubleshoot a deployment,
since there is no reference file for the network config
which was applied. This change renames the file and
appends the date_time to the file name to preserve the file
for troubleshooting.

Closes-bug: 1816920
Change-Id: I79a6c5428958c1508eb16f91096cb96d7ddc03f6