Add OvnDbInternal to EndpointMap and use it for ovn_db_host
OVN controller/metadata use ovn_dbs_vip hiera key to configure
the central ovn DB. This key is not available on split control
plane or multi cell setup and therefore installation fails.

With this change a new entry gets created in the EndpointMap
named OvnDbInternal. This can then be exported for an overcloud
stack and can be used as an input for the cell stack.

The information from the EndpointMap is used for ovn-metadata
and ovn-controller as the ovn_db_host information in puppet-tripleo

Change-Id: I0d9eb663405d1113ea84e3c12651a3f0dbdfc75d
Closes-Bug: #1817524