container-image-prepare: redirect all output to logfile
If container_image_prepare_debug, we also want to redirect all the
output into the same logfile used to write image operations.

Indeed, some command output isn't shown in the logfile, even with
--debug, but sent to sdterr.

To make it easier for operators on where to find logs, let's just use
the logfile in place and export everything in there, only when debug is
activated.

Change-Id: I66e846bda2bef0d27a2a26a4083249c583982e0c