[Rocky-only] Install and configure tmpwatch for log cleanup
Installing and configuring tmpwatch allows to get rid of some
ugly things in logrotate configuration. As the container has no
network access anymore, we have to install the tool on the host
directly - this isn't that bad.

In order to avoid issues with logrotate manage logs, we explicitely
exclude patterns manage in the specific logorate configuration.

Also, always in order to avoid issues and ensure logrotate does its
own cleanup, we clean files one day later.

Note: this patch is the port of [1], without the "deployment" directory.
We can't really cherry-pick things since the tree is modified in master.

[1] https://review.openstack.org/641608

Change-Id: I5c59f337f2a62bc44e459f13de9bab6b4ab1a9f5