Make sure ContainerCli is set to the stack
ContainerCli has been used everwhere now to differentiate a deployment
on Docker and on Podman. The parameter needs to be in the stack anyway,
so let's have it in overcloud-resource-registry-puppet.

Note: default to Podman, except when Pacemaker is deployed where we
don't support Podman yet on CentOS7. It will change with CentOS8.

Change-Id: I02aa484cabfb851f80ddd90ddefdfe2d5cfbb573
Closes-Bug: #1819449