Reload rsyslog/cron when we change timezones
If the default timezone in the overcloud image or deployed system is
changed on install, rsyslog will continue to use that timezone until
it's reloaded. Do a reload of rsyslog if we update the timezone file.
The same happens with crond so we need to reload that again.

Change-Id: I7cfde5963b4fce010e936b4fd5bbe96a39cc5c17
Closes-Bug: #1819548
(cherry picked from commit 2182e0b0d0c07cd0a90f8ebb783072e9ee41e81e)