Switch scenario004-multinode-containers to Ceph filestore
With [1] we now have both LVM and /dev/loop3 for either
bluestore or filestore in the same branch. Let's include
filestore testing in queens by switching 004 to filestore
but keep 001 on bluestore.

Patching Queens directly to test both blue and file store.
I expect Rocky and newer will go out with bluestore so not
switching those.

[1] https://review.openstack.org/#/c/641479

Change-Id: I13c6c3fc88136157b8036c15d23589380b1fcdd8