Run chown for nova log files on every run to fix wrong permissions
If nova-manage command was triggered on a host for the first time as root
(usually manual runs) the nova-manage.log gets created as root user. On
overcloud deploy runs the nova-manage command is run as nova user. In such
situation the overcloud deploy fails as the nova user can not write to the
nova-manage.log. With this change we run the chown of the logs files on
every overcloud deploy to fix the nova-manage.log file permissions.

Closes-Bug: #1820590

Change-Id: Iaa8db09712da6c0d9553fab39d7d5b50fa7cf287