Refactor ovn_dbs upgrade_tasks.
Reorder the upgrade_tasks for the ovn_dbs service in a similar way
it's being done for the rest of pacemaker managed services. Also,
ensure indempotency in the case the cluster is down as it will occur
for the S to R upgrades workflow.

Partial-Bug: #1831690
Change-Id: I8fe302607544653032a011f1c932647adec5bfa0
(cherry picked from commit 05f4b253b84526850959f1b827a261b7573c006c)