Remove UndercloudCtlplaneSubnets defaults in UC env
The defaults in environments/undercloud.yaml will be
merged with what the was defined in undercloud.conf.
If the 'ctlplane-subnet' is not used in undercloud.conf
the default is merged, and we end up with an additional
subnet that was not requested by the undercloud installed.

This change set the default to {} in:
  environments/undercloud.yaml

Also add's a comment to hightlight that there should be
a empty dict default.

Closes-Bug: #1820330
Change-Id: Idc17154b2e1e143ba4e84b9b96f5355fc9230172
(cherry picked from commit ea80847f9697f6b02f9cbaedcafc909e37ce5161)