Fixes for freeipa_setup.sh
We make the following changes in this review:
a) We only remove openstack-dashboard if it exists
   as the script fails otherwise

b) We differentiate between the set of packages that we install
   when using CentOS/RHEL7 and when using Centos/RHEL8
   This condition is needed because some packages do either not
   exist in RHEL8 or have changed names entirely.

c) We do not enable haveged unconditionally since it does not
   exist in rhel8 and we also do not remove ssl.conf unconditionally
   on rhel8 as that breaks things as well.

It is expected that the list for RHEL8 might change over time as
things get added (e.g. EPEL).

Change-Id: I10e082c0bc75677115364c055209344d1b094fec