Add OS::TripleO::NeutronAZConfig
This service is mapped to OS::Heat::None by default, but when enabled,
it allows to configure defaut AZ for neutron, AZ for L3/DHCP agents and
AZ-aware network scheduling/routing. The zone name is the stack
name by default, but can be overridden with parameter values.

Change-Id: Ie5762c61cd615d7c115069381d6c43c468806e47
implements: blueprint split-controlplane-templates