Add OS::TripleO::NovaAZConfig
This service is mapped to OS::Heat::None by default, but when enabled,
it uses external deploy tasks to create a new aggregate and zone in
Nova. All Compute nodes in the stack will be added to the zone by
default.

The zone name is the stack name by default, but can be overridden with
parameter values.

Change-Id: I0afb2265949f1222a86b350734e62da673e83aa9
Depends-On:I064f872e20537642880d5482cb9e0c2a801715ef
implements: blueprint split-controlplane-templates