Add CinderEtcdLocalConnect parameter
The parameter, when set to true, will configure cinder-volume to connect
to Etcd through the node's own local IP on the Etcd network, instead of
a VIP on the network.

This is useful when deploying cinder-volume in an A/A configuration at
an edge site with the HCI roles. As Etcd and cinder-volume are both
running on the same node (typically 3 nodes configured the same), then
each node can just connect directly to Etcd without having to go through
a VIP. Additionally, we have no VIP management at the edge sites
presently.

Change-Id: I8a8825ecff9fc99b5de7390075470356397d85a2
implements: blueprint split-controlplane-templates