Fix reload notification file
We want to trigger the reload on the localtime file (/etc/localtime) not
the actual zone file because that is the symlink that gets updated when
the timezone is changed.

NOTE: This is <= Rocky only because we switched to ansible managment in
Stein.

Change-Id: I2bdafeb59c3feb86a30a46eeb803bc9795c9c820
Closes-Bug: #1819548
(cherry picked from commit ac4bd32a70db90f55100bb74f939dad99547cde3)