Deprecate nova-consoleauth service
As of Rocky [1], the nova-consoleauth service has been deprecated and
cell databases are used for storing token authorizations. All new consoles
will be supported by the database backend and existing consoles will be
reset. Console proxies must be run per cell because the new console token
authorizations are stored in cell databases.

Lets deprecate it also in tripleo that it can be removed in a later
release.

[1] https://docs.openstack.org/releasenotes/nova/rocky.html

Change-Id: I68485a6c4da4476d07ec0ab5e7b5a4c528820a4f