Stop writing /var/lib/container-startup-configs.json
Now that we've dropped docker-toool we no longer have use
for the /var/lib/container-startup-configs.json file to be laid
down on disk.

As part of this change we now check for the step 1 tasks instead
of the combined startup configs when determining whether to
re-run the common startup ansible tasks.

Change-Id: I3c50d8364823073341b5f39ecce20a512e4a82e1