Do not archive ceph-ansible fetch directory if it is empty
There are conditions, e.g. external ceph deployment, where
ceph-ansible will not produce a fetch directory so we should
not always assume there is something to archive.

Add a task to register the contents of the ceph-ansible fetch
directory and only archive that directory if the contents exist.

Also, remove the '--no-overwrite-dir' option from tar as it
seems to have broken repeat deployments.

Change-Id: I56eae30d73bf5cc6b1ae413c6841f86c2a4e86b5
Closes-Bug: #1823759
Closes-Bug: #1823229