Increase default ulimit values for Neutron agents containers
For containers where Neutron DHCP, L3 and OVS agents are running
there is limit for number of opened files set.
This limit was configured by default to 1024 which looks that is
quite low and can cause problems even with relatively small number
of networks or routers handled by agent.

This patch increase default ulimit values for those containers to be
16x higher than it was before. Such value should allow to handle much
more routers or networks and shouldn't be too high for any system.

Change-Id: Ibc8fa0c1ccd4f60ec7c876623a7262003197f67d
Closes-bug: #1824020