Add {{network.name}}NetworkVlanID to THT/net-config-*
With the default network_data.yaml and default roles_data.yaml
all composable networks are included. When deploying plain
defaults without any network-isolation, i.e ctlplane network
only the jinja2 rendered templates still include all the nets.

With change: https://review.openstack.org/645159 the vlan
parameter is passed to NIC configs for all networks enabled
for the role in roles data if networks data define a vlan id.

Because of this we need to add the parameter to the nic config
template unless networks data and roles data are customized.
(i.e networks that are'nt actually used are removed from the
roles data and networks data.)

Closes-Bug: #1823997
Change-Id: Ic5fe9c8097de814ce3ad7032f42b31859051f70a