Correct error catching while in "--check" mode
When we're in "--check" mode with Ansible, the variable is defined,
but its content doesn't match the one we might get when actually applying
the run.
The right way to detect that is to check if the variable is changed.

In "--check", it is not changed, meaning we can't get the "rc" attribute.

I5851dc7820fdcc4f5790980d94b81622ce3b0c8d didn't do the right check, and
does not improve the situation.

Change-Id: I77c655addd0fe01948584fb43ba24d5b24ff330e
Related-Bug: #1823841