Make node_ips generic in the Heat template
Replaces the expressions for the node_ips values which resulted in
hardcoded data coming out of the Heat stack with a generic Ansible
jinja expression that will read the same values from the Ansible
inventory and facts.

The data will be rendered at config-download runtime instead of
during the Heat stack operation. This allows for more generic use of
the config-download playbooks with the Ansible inventory alone.

For example, adding a new node to the Ansible inventory and adding
it to the inventory group for a service would cause that node's ip
to be automatically added to the corresponding node_ips lists in the
hieradata. No Heat stack update would be required to add the node.

Depends-On: Iae9ccdf6fad2fabdce66650ee7b064f1ac9528e4
Change-Id: I2a5a7a8ecc992ce558301abfb7a7a098b22743fe
implements: blueprint reduce-deployment-resources